Lots of comments and cleanups
