+++ CanvasBlittingTool/CBTJSONCoordinator.swift
@@ -10,7 +10,7 @@ import Cocoa
 
 class CBTJSONCoordinator: CBTCoordinator
 {
-    var data:[[[Int]]] = [[[]]];
+    var data:[[[Int]]];
     
     init (frames: Int)
     {
@@ -18,6 +18,7 @@ class CBTJSONCoordinator: CBTCoordinator
         super.init();
     }
     
+    //Adds one cell write to our manifest data
     func addCellData(sourceIndex:Int, destinationIndex:Int, frame: Int, callbackClosure: () -> Void)
     {
         var operation = NSBlockOperation { () -> Void in
